Allow to handle errors by users
If something went wrong with the upload I'd like to handle it in a Catch-node.

According to [knolleary: Node.error() not caught in Catch node](https://discourse.nodered.org/t/node-error-not-caught-in-catch-node/1301/3) this can be achieved by passing the original `msg`-object to the node.error() method.

Let me know if this is something you'd like to integrate?

Btw. very good work. I use you nodes very often in my flows 👍  Thank you for creating this nextcloud integration.